Update PHPUnit configuration files
This change replaces the existing PHPUnit configuration file with a new
one generated by PHPUnit, instead of borrowed and slightly modified from
another project.

The motivation for generating a new config file was to address a problem
when running the test suite with code coverage enabled. The best
resolution that I found was https://stackoverflow.com/a/63941242/222011
from Sebastian Bergmann, which worked for me.