Load the WebhookValidationMiddleware with a custom factory
This change registers the WebhookValidationMiddleware as a service with
the DI container, adding a factory class to provide it when requested.

Because of the new nature of how the class is loaded, the class has been
refactored to retrieve the Twilio Auth Token from the application's
config, inside the 'twilio' namespace within that config, rather than
via an appropriately named environment variable as before. Doing so
brings it in line with how other classes and services are configured.