Set the site URL
This change sets the Antora site URL which I believe is required for the
Antora-generated redirects file to be generated correctly. In short, it
needs the relative URL to be set, so that it is included in all of the
generated redirects. That way, they'll work correctly when the site is
deployed on GitLab.

See https://docs.antora.org/antora/latest/playbook/site-url/ and
https://docs.gitlab.com/user/project/pages/redirects/ for more
information.